+++ src/lightsclient/InputReceiver.java
@@ -53,7 +53,7 @@ public class InputReceiver implements Receiver {
 						sendData(
 								new MyMessage(part.getChannel(), Type.TIME_UPDATE, part.getTime(), part.getNextTime()));
 					} else {
-						Float time = part.isPossible(noteNum);
+						Double time = part.isPossible(noteNum);
 
 						if (time != null) {
 							// sendData(new MyMessage(part.getChannel(),
@@ -78,7 +78,7 @@ public class InputReceiver implements Receiver {
 	}
 
 	// notify with MyMessage
-	public Float notify(MyMessage message) {
+	public Double notify(MyMessage message) {
 		// System.out.println(Thread.currentThread().getName());
 
 		// parse message
@@ -108,7 +108,7 @@ public class InputReceiver implements Receiver {
 			break;
 		case TIME_UPDATE:
 			// figure out what to do based upon channel
-			Float newTime = null;
+			Double newTime = null;
 			switch (channel) {
 			case 2: // next part
 				System.out.println("Calling nextPart() from " + part.getChannel());
@@ -149,7 +149,7 @@ public class InputReceiver implements Receiver {
 		// }
 	}
 
-	public void changeTime(float newTime) {
+	public void changeTime(double newTime) {
 		// if part time or measure time != new time, change part
 		if ((part.getPartTime() != newTime) || (part.getMeasureTime() != newTime)) {
 			part.changeTime(newTime);
